moved useQuesty to hooks directory. Refactored AddChallenge.tsx to useCheckRole hook
