+++ client/src/components/NavigationBar.tsx
@@ -6,7 +6,7 @@ import { NAV_CONFIG } from "../App.tsx";
 import NewUserPopup from './NewUserPopup'; // Make sure this path is correct
 import { QuestionCircle } from "react-bootstrap-icons";
 import { UserData } from '../types/UserData';
-import { useQuery } from '../helpers/useQuery.tsx';
+import { useQuery } from '../hooks/useQuery.tsx';
 
 function NavigationBar() {
   const [showNewUserPopup, setShowNewUserPopup] = useState(false);
