--- client/src/pages/Landing.tsx
@@ -1,18 +1,0 @@
-// import { DrawIoEmbed } from "react-drawio";
-// import { useState } from "react";
-// import { Button } from "react-bootstrap";
-
-function Landing() {
-
-  // const [data, setData] = useState<string | null>(null);
-
-
-  return <>
-          <h1>UML Mentor</h1>
-          {/* <DrawIoEmbed onSave={(data) => {setData(data.xml)}}></DrawIoEmbed>
-          <Button onClick={() => {console.log(data?.toString())}}>Print data</Button> */}
-        </>;
-}
-
-export default Landing;
-
